#516 totp certification

+++ app/controllers/ip_lists_controller.rb
@@ -11,10 +11,10 @@ class IpListsController < ApplicationController
       return to_error_path
     end
 
-    return to_error_path unless user.linable?
+    return to_error_path unless user.linable? || user.otp_enabled
 
     ip = IpList.white_ip!(request.remote_ip, user)
-    if LineHookService.push_message(user.line_id, I18n.t('line_authentication', token: ip.token)).is_a?(Net::HTTPSuccess)
+    if user.otp_enabled || LineHookService.push_message(user.line_id, I18n.t('line_authentication', token: ip.token)).is_a?(Net::HTTPSuccess)
       redirect_to edit_ip_list_path(ip)
     else
       ip.destroy
@@ -25,13 +25,14 @@ class IpListsController < ApplicationController
   def edit; end
 
   def update
-    if @ip.authenticate?(params[:token])
-      @ip.updated_expired_on
-      log_in(@ip.created_user)
-      redirect_to menu_index_path
-    else
-      redirect_to new_ip_list_path
+    if @ip.created_user.otp_enabled
+      return redirect_to new_ip_list_path unless @ip.created_user.totp.verify(params[:token], drift_behind: 30, drift_ahead: 30)
+    elsif !@ip.authenticate?(params[:token])
+      return redirect_to new_ip_list_path
     end
+    @ip.updated_expired_on
+    log_in(@ip.created_user)
+    redirect_to menu_index_path
   end
 
   private
